Document health, metrics & OpenTracing extensions
