+++ tooling/maven/package-maven-plugin/src/main/java/org/apache/camel/quarkus/maven/model/ComponentModel.java
@@ -16,37 +16,14 @@
  */
 package org.apache.camel.quarkus.maven.model;
 
-import org.apache.camel.quarkus.maven.StringHelper;
+public class ComponentModel extends AbstractModel {
 
-import static org.apache.camel.quarkus.maven.StringHelper.cutLastZeroDigit;
-
-public class ComponentModel {
-
-    private String kind;
     private String scheme;
     private String syntax;
     private String alternativeSyntax;
     private String alternativeSchemes;
-    private String title;
-    private String description;
-    private String firstVersion;
-    private String label;
-    private String deprecated;
-    private String deprecationNote;
     private String consumerOnly;
     private String producerOnly;
-    private String javaType;
-    private String groupId;
-    private String artifactId;
-    private String version;
-
-    public String getKind() {
-        return kind;
-    }
-
-    public void setKind(String kind) {
-        this.kind = kind;
-    }
 
     public String getScheme() {
         return scheme;
@@ -80,54 +57,6 @@ public class ComponentModel {
         this.alternativeSchemes = alternativeSchemes;
     }
 
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getFirstVersion() {
-        return firstVersion;
-    }
-
-    public void setFirstVersion(String firstVersion) {
-        this.firstVersion = firstVersion;
-    }
-
-    public String getLabel() {
-        return label;
-    }
-
-    public void setLabel(String label) {
-        this.label = label;
-    }
-
-    public String getDeprecated() {
-        return deprecated;
-    }
-
-    public void setDeprecated(String deprecated) {
-        this.deprecated = deprecated;
-    }
-
-    public String getDeprecationNote() {
-        return deprecationNote;
-    }
-
-    public void setDeprecationNote(String deprecationNote) {
-        this.deprecationNote = deprecationNote;
-    }
-
     public String getConsumerOnly() {
         return consumerOnly;
     }
@@ -144,47 +73,14 @@ public class ComponentModel {
         this.producerOnly = producerOnly;
     }
 
-    public String getJavaType() {
-        return javaType;
+    @Override
+    String getDocLinkSection() {
+        return "components";
     }
 
-    public void setJavaType(String javaType) {
-        this.javaType = javaType;
+    @Override
+    String getDocLinkDocument() {
+        return scheme + "-component.html";
     }
 
-    public String getGroupId() {
-        return groupId;
-    }
-
-    public void setGroupId(String groupId) {
-        this.groupId = groupId;
-    }
-
-    public String getArtifactId() {
-        return artifactId;
-    }
-
-    public void setArtifactId(String artifactId) {
-        this.artifactId = artifactId;
-    }
-
-    public String getVersion() {
-        return version;
-    }
-
-    public void setVersion(String version) {
-        this.version = version;
-    }
-
-    public String getShortJavaType() {
-        return StringHelper.getClassShortName(javaType);
-    }
-
-    public String getDocLink() {
-        return "https://camel.apache.org/components/latest/" + scheme + "-component.html";
-    }
-
-    public String getFirstVersionShort() {
-        return cutLastZeroDigit(firstVersion);
-    }
 }
