+++ tooling/maven/package-maven-plugin/src/main/java/org/apache/camel/quarkus/maven/model/OtherModel.java
@@ -16,130 +16,19 @@
  */
 package org.apache.camel.quarkus.maven.model;
 
-import org.apache.camel.quarkus.maven.StringHelper;
-
-import static org.apache.camel.quarkus.maven.StringHelper.cutLastZeroDigit;
-
-public class OtherModel {
-
-    private String kind;
-    private String name;
-    private String title;
-    private String description;
-    private String firstVersion;
-    private String label;
-    private String deprecated;
-    private String deprecationNote;
-    private String groupId;
-    private String artifactId;
-    private String version;
-    private String javaType;
-
-    public String getJavaType() {
-        return javaType;
-    }
-
-    public void setJavaType(String javaType) {
-        this.javaType = javaType;
-    }
-
-    public String getKind() {
-        return kind;
-    }
-
-    public void setKind(String kind) {
-        this.kind = kind;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getFirstVersion() {
-        return firstVersion;
-    }
-
-    public void setFirstVersion(String firstVersion) {
-        this.firstVersion = firstVersion;
-    }
-
-    public String getLabel() {
-        return label;
-    }
-
-    public void setLabel(String label) {
-        this.label = label;
-    }
-
-    public String getDeprecated() {
-        return deprecated;
-    }
-
-    public void setDeprecated(String deprecated) {
-        this.deprecated = deprecated;
-    }
-
-    public String getDeprecationNote() {
-        return deprecationNote;
-    }
-
-    public void setDeprecationNote(String deprecationNote) {
-        this.deprecationNote = deprecationNote;
-    }
-
-    public String getGroupId() {
-        return groupId;
-    }
-
-    public void setGroupId(String groupId) {
-        this.groupId = groupId;
-    }
-
-    public String getArtifactId() {
-        return artifactId;
-    }
-
-    public void setArtifactId(String artifactId) {
-        this.artifactId = artifactId;
-    }
-
-    public String getVersion() {
-        return version;
-    }
-
-    public void setVersion(String version) {
-        this.version = version;
-    }
+public class OtherModel extends AbstractModel {
 
     public String getDocLink() {
-        return "https://camel.apache.org/components/latest/" + name + ".html";
+        return localDocExists() ? getLocalDocLink() : null;
     }
 
-    public String getFirstVersionShort() {
-        return cutLastZeroDigit(firstVersion);
+    @Override
+    String getDocLinkSection() {
+        return null;
     }
 
-    public String getShortJavaType() {
-        return StringHelper.getClassShortName(javaType);
+    @Override
+    String getDocLinkDocument() {
+        return null;
     }
 }
