+++ tooling/maven/package-maven-plugin/src/main/resources/readme-components.mvel
@@ -6,7 +6,7 @@ Number of Camel components: @{components.size} in @{numberOfArtifacts} JAR artif
 |===
 | Component | Since | Description
 @foreach{row : components}
-| link:@{row.docLink}[@{row.title}] (@{row.artifactId}) +
+| @{row.docLink}[@{row.title}] (@{row.artifactId}) +
 `@{row.syntax}` | @{row.firstVersionShort} | @if{row.deprecated == "true"}*deprecated* @end{}@{util.escape(row.description)}
 @end{}
 |===
